Split v1 and v2 tests
