+++ datadog_checks_base/tests/base/checks/openmetrics/test_legacy/test_compat_scraper.py
