+++ datadog_checks_base/tests/base/checks/openmetrics/test_legacy/test_openmetrics.py
@@ -23,7 +23,7 @@ from datadog_checks.dev import get_here
 from datadog_checks.dev.http import MockResponse
 
 text_content_type = 'text/plain; version=0.0.4'
-FIXTURE_PATH = os.path.abspath(os.path.join(get_here(), '..', '..', '..', 'fixtures', 'prometheus'))
+FIXTURE_PATH = os.path.abspath(os.path.join(get_here(), '..', '..', '..', '..', 'fixtures', 'prometheus'))
 
 FAKE_ENDPOINT = 'http://fake.endpoint:10055/metrics'
