+++ datadog_checks_base/tests/base/checks/openmetrics/test_legacy/test_openmetrics_base_check.py
@@ -5,7 +5,7 @@ from mock import patch
 
 from datadog_checks.checks.openmetrics import OpenMetricsBaseCheck
 
-FIXTURE_PATH = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', '..', 'fixtures', 'bearer_tokens')
+FIXTURE_PATH = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', '..', '..', 'fixtures', 'bearer_tokens')
 
 
 class TestSignature:
