+++ datadog_checks_base/tests/base/checks/openmetrics/test_v2/test_transformers/test_counter.py
