+++ datadog_checks_base/tests/base/checks/openmetrics/test_v2/utils.py
@@ -3,24 +3,9 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from datadog_checks.base import OpenMetricsBaseCheckV2
 
-# TODO: remove `try` when we drop Python 2
-try:
-    from datadog_checks.base.checks.openmetrics.v2.scraper import OpenMetricsCompatibilityScraper
-
-    class LegacyCheck(OpenMetricsBaseCheckV2):
-        def create_scraper(self, config):
-            return OpenMetricsCompatibilityScraper(self, self.get_config_with_defaults(config))
-
-except Exception:
-    pass
-
 
 def get_check(instance=None, init_config=None):
     return _get_check(OpenMetricsBaseCheckV2, instance, init_config)
-
-
-def get_legacy_check(instance=None, init_config=None):
-    return _get_check(LegacyCheck, instance, init_config)
 
 
 def _get_check(cls, instance, init_config):
